+++ test/libsolidity/syntaxTests/parsing/array_range_no_start.sol
@@ -1,0 +1,7 @@
+contract C {
+    function f(uint256[] calldata x) external pure {
+        x[:];
+    }
+}
+// ----
+// ParserError: (76-77): Expected primary expression.